Version 1.0
This application creates dfa and then tests wether certain inputs are in the language recognised by it or not.

How it creates is of a format provided in the readme file.
you give it the dfa states and then provide a string of 0s and 1s and it tells wether it is in the language or not,